Limit cat queries to user colonies

+++ backend/service/src/common/base-crud.service.ts
@@ -41,13 +41,15 @@ export class BaseCrudService<T extends BaseEntity> implements ICrudService<T> {
   }
 
   async update(id: number, updateInput: Record<string, any>): Promise<T> {
-    const entity = await this.repository.findOne(id);
+    const entity = await this.findOne(id);
     if (!entity) return;
     Object.assign(entity, updateInput);
     return entity.save();
   }
 
   async remove(id: number): Promise<boolean> {
+    const entity = await this.findOne(id);
+    if (!entity) return false;
     const result = await this.repository.delete(id);
     return result.affected > 0;
   }
